requirements.txt
alpaca-py==0.31.0
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
msgpack==1.1.0
numpy==2.1.2
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sseclient-py==1.8.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websockets==13.1
